+++ woocommerce-gateway-placetopay.php
@@ -6,14 +6,14 @@
  * Author:      PlacetoPay
  * Author URI:  https://www.placetopay.com/
  * Developer:   Cristian Salazar
- * Version:     2.13.1
+ * Version:     2.16.0
  *
  *
  * @package PlacetoPay/WC_Gateway_PlacetoPay
  *
  * @author Soporte <soporte@placetopay.com>
  * @copyright (c) 2013-2017 EGM Ingenieria sin fronteras S.A.S.
- * @version 2.13.1
+ * @version 2.16.0
  */
 
 if (!defined('ABSPATH')) {
@@ -67,7 +67,7 @@ function wc_gateway_placetopay()
     }
 
     require_once(__DIR__ . '/vendor/autoload.php');
-    return \PlacetoPay\PaymentMethod\WC_Gateway_PlacetoPay::getInstance('2.13.1', __FILE__);
+    return \PlacetoPay\PaymentMethod\WC_Gateway_PlacetoPay::getInstance('2.16.0', __FILE__);
 }
 
 add_action('plugins_loaded', 'wc_gateway_placetopay', 0);